fix(channel): address Mattermost review gaps
