+++ packages/desktop/src/renderer/components/settings/SettingsModal/contents/channels/platforms/DiscordConfigForm.tsx
@@ -20,7 +20,7 @@ import {
   buildChannelAssistantBinding,
   getDefaultChannelAssistant,
   resolveChannelAssistantSelection,
-} from './assistantBinding';
+} from '../assistantBinding';
 
 /**
  * Preference row component
@@ -53,14 +53,14 @@ const SectionHeader: React.FC<{ title: string; action?: React.ReactNode }> = ({ 
   </div>
 );
 
-interface SlackConfigFormProps {
+interface DiscordConfigFormProps {
   pluginStatus: IChannelPluginStatus | null;
   modelSelection: GoogleModelSelection;
   onStatusChange: (status: IChannelPluginStatus | null) => void;
   onTokenChange?: (token: string) => void;
 }
 
-const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
+const DiscordConfigForm: React.FC<DiscordConfigFormProps> = ({
   pluginStatus,
   modelSelection,
   onStatusChange,
@@ -70,7 +70,6 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
   const localeKey = resolveLocaleKey(i18n?.language ?? 'en-US');
 
   const [botToken, setBotToken] = useState('');
-  const [appToken, setAppToken] = useState('');
   const [testLoading, setTestLoading] = useState(false);
   const [tokenTested, setTokenTested] = useState(false);
   const [testedBotUsername, setTestedBotUsername] = useState<string | null>(null);
@@ -79,7 +78,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
   const [pendingPairings, setPendingPairings] = useState<IChannelPairingRequest[]>([]);
   const [authorizedUsers, setAuthorizedUsers] = useState<IChannelUser[]>([]);
 
-  // Assistant selection (used for Slack conversations)
+  // Assistant selection (used for Discord conversations)
   const [availableAssistants, setAvailableAssistants] = useState<Assistant[]>([]);
   const [selectedAssistant, setSelectedAssistant] = useState<Assistant | null>(null);
   const [hasBrokenSavedAssistant, setHasBrokenSavedAssistant] = useState(false);
@@ -90,7 +89,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
     try {
       const pairings = await channel.getPendingPairings.invoke();
       if (pairings) {
-        setPendingPairings(pairings.filter((p) => p.platformType === 'slack'));
+        setPendingPairings(pairings.filter((p) => p.platformType === 'discord'));
       }
     } catch (error) {
       console.error('[ChannelSettings] Failed to load pending pairings:', error);
@@ -105,7 +104,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
     try {
       const users = await channel.getAuthorizedUsers.invoke();
       if (users) {
-        setAuthorizedUsers(users.filter((u) => u.platformType === 'slack'));
+        setAuthorizedUsers(users.filter((u) => u.platformType === 'discord'));
       }
     } catch (error) {
       console.error('[ChannelSettings] Failed to load authorized users:', error);
@@ -126,7 +125,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
       try {
         const [assistantList, saved] = await Promise.all([
           assistants.list.invoke(),
-          channel.getPlatformSettings.invoke({ platform: 'slack' }),
+          channel.getPlatformSettings.invoke({ platform: 'discord' }),
         ]);
 
         setAvailableAssistants(assistantList);
@@ -140,7 +139,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
         setHasBrokenSavedAssistant(selection.hasBrokenSavedAssistant);
         setSelectedAssistant(nextAssistant);
       } catch (error) {
-        console.error('[SlackConfig] Failed to load assistants:', error);
+        console.error('[DiscordConfig] Failed to load assistants:', error);
       }
     };
 
@@ -150,12 +149,12 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
   const persistSelectedAssistant = async (assistant: Assistant) => {
     try {
       await channel.setAssistantSetting.invoke({
-        platform: 'slack',
+        platform: 'discord',
         assistant: buildChannelAssistantBinding(assistant),
       });
       Message.success(t('settings.assistant.agentSwitched', 'Assistant switched successfully'));
     } catch (error) {
-      console.error('[SlackConfig] Failed to save assistant:', error);
+      console.error('[DiscordConfig] Failed to save assistant:', error);
       Message.error(t('common.saveFailed', 'Failed to save'));
     }
   };
@@ -163,7 +162,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
   // Listen for pairing requests
   useEffect(() => {
     const unsubscribe = channel.pairingRequested.on((request) => {
-      if (request.platformType !== 'slack') return;
+      if (request.platformType !== 'discord') return;
       setPendingPairings((prev) => {
         const exists = prev.some((p) => p.code === request.code);
         if (exists) return prev;
@@ -186,26 +185,12 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
     return () => unsubscribe();
   }, []);
 
-  // Test Slack connection (only validates the Bot Token via auth.test)
+  // Test Discord connection (validates the Bot Token via GET /users/@me)
   const handleTestConnection = async () => {
-    const trimmedBotToken = botToken.trim();
-    const trimmedAppToken = appToken.trim();
+    const trimmedToken = botToken.trim();
 
-    if (!trimmedBotToken) {
-      Message.warning(t('settings.assistant.slackBotTokenRequired', 'Please enter a Bot Token (xoxb-)'));
-      return;
-    }
-    if (!trimmedAppToken) {
-      Message.warning(t('settings.assistant.slackAppTokenRequired', 'Please enter an App-Level Token (xapp-)'));
-      return;
-    }
-    // Soft prefix validation (hard validation happens on the backend)
-    if (!trimmedBotToken.startsWith('xoxb-')) {
-      Message.warning(t('settings.assistant.slackBotTokenPrefix', 'Bot Token should start with "xoxb-"'));
-      return;
-    }
-    if (!trimmedAppToken.startsWith('xapp-')) {
-      Message.warning(t('settings.assistant.slackAppTokenPrefix', 'App-Level Token should start with "xapp-"'));
+    if (!trimmedToken) {
+      Message.warning(t('settings.assistant.tokenRequired', 'Please enter a bot token'));
       return;
     }
 
@@ -213,12 +198,10 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
     setTokenTested(false);
     setTestedBotUsername(null);
     try {
-      // testPlugin returns { success, bot_username?, error? } directly.
-      // Slack test only validates the Bot Token (auth.test); the App-Level
-      // Token is only needed when starting the Socket Mode connection.
+      // testPlugin returns { success, bot_username?, error? } directly
       const result = await channel.testPlugin.invoke({
-        plugin_id: 'slack',
-        token: trimmedBotToken,
+        plugin_id: 'discord',
+        token: trimmedToken,
       });
 
       if (result.success) {
@@ -245,44 +228,32 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
     }
   };
 
-  // Auto-enable plugin after successful test (submits both tokens)
+  // Auto-enable plugin after successful test
   const handleAutoEnable = async () => {
     try {
       // enablePlugin returns void; success if no throw
       await channel.enablePlugin.invoke({
-        plugin_id: 'slack',
-        config: {
-          credentials: {
-            token: botToken.trim(),
-            app_token: appToken.trim(),
-          },
-        },
+        plugin_id: 'discord',
+        config: { credentials: { token: botToken.trim() } },
       });
 
-      Message.success(t('settings.assistant.slackPluginEnabled', 'Slack bot enabled'));
+      Message.success(t('settings.assistant.discordPluginEnabled', 'Discord bot enabled'));
       const plugins = await channel.getPluginStatus.invoke();
       if (plugins) {
-        const slackPlugin = plugins.find((p) => p.type === 'slack');
-        onStatusChange(slackPlugin || null);
+        const discordPlugin = plugins.find((p) => p.type === 'discord');
+        onStatusChange(discordPlugin || null);
       }
     } catch (error: unknown) {
       console.error('[ChannelSettings] Auto-enable failed:', error);
     }
   };
 
-  // Reset token tested state when the bot token changes
-  const handleBotTokenChange = (value: string) => {
+  // Reset token tested state when the token changes
+  const handleTokenChange = (value: string) => {
     setBotToken(value);
     setTokenTested(false);
     setTestedBotUsername(null);
     onTokenChange?.(value);
-  };
-
-  // Reset token tested state when the app token changes
-  const handleAppTokenChange = (value: string) => {
-    setAppToken(value);
-    setTokenTested(false);
-    setTestedBotUsername(null);
   };
 
   // Approve pairing
@@ -341,59 +312,21 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
     ? resolveAssistantName(selectedAssistant, localeKey, selectedAssistant.name)
     : t('settings.assistant.name', 'Assistant');
 
-  // Tokens are locked once there are authorized users (mirrors Telegram).
-  const tokensLocked = authorizedUsers.length > 0;
+  // Token is locked once there are authorized users (mirrors Telegram/Slack).
+  const tokenLocked = authorizedUsers.length > 0;
 
   return (
     <div className='flex flex-col gap-24px'>
-      {/* Bot Token (xoxb-) */}
+      {/* Bot Token */}
       <PreferenceRow
-        label={t('settings.assistant.slackBotToken', 'Bot Token')}
+        label={t('settings.assistant.discordBotToken', 'Bot Token')}
         description={t(
-          'settings.assistant.slackBotTokenDesc',
-          'Create a Slack app, install it to your workspace, then copy the Bot User OAuth Token (xoxb-).'
-        )}
-      >
-        {tokensLocked ? (
-          <Tooltip
-            content={t(
-              'settings.assistant.tokenLocked',
-              'Please close the Channel and delete all authorized users before modifying the configuration'
-            )}
-          >
-            <span>
-              <Input.Password
-                value={botToken}
-                onChange={handleBotTokenChange}
-                placeholder={tokensLocked || pluginStatus?.hasToken ? '••••••••••••••••' : 'xoxb-...'}
-                style={{ width: 260 }}
-                visibilityToggle
-                disabled={tokensLocked}
-              />
-            </span>
-          </Tooltip>
-        ) : (
-          <Input.Password
-            value={botToken}
-            onChange={handleBotTokenChange}
-            placeholder={tokensLocked || pluginStatus?.hasToken ? '••••••••••••••••' : 'xoxb-...'}
-            style={{ width: 260 }}
-            visibilityToggle
-            disabled={tokensLocked}
-          />
-        )}
-      </PreferenceRow>
-
-      {/* App-Level Token (xapp-) */}
-      <PreferenceRow
-        label={t('settings.assistant.slackAppToken', 'App-Level Token')}
-        description={t(
-          'settings.assistant.slackAppTokenDesc',
-          'Enable Socket Mode and generate an App-Level Token with the connections:write scope (xapp-).'
+          'settings.assistant.discordBotTokenDesc',
+          'Create a Discord application, add a Bot, enable the MESSAGE CONTENT intent, and copy the Bot Token.'
         )}
       >
         <div className='flex items-center gap-8px'>
-          {tokensLocked ? (
+          {tokenLocked ? (
             <Tooltip
               content={t(
                 'settings.assistant.tokenLocked',
@@ -402,26 +335,26 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
             >
               <span>
                 <Input.Password
-                  value={appToken}
-                  onChange={handleAppTokenChange}
-                  placeholder={tokensLocked || pluginStatus?.hasToken ? '••••••••••••••••' : 'xapp-...'}
+                  value={botToken}
+                  onChange={handleTokenChange}
+                  placeholder={tokenLocked || pluginStatus?.hasToken ? '••••••••••••••••' : 'MTAx...'}
                   style={{ width: 260 }}
                   visibilityToggle
-                  disabled={tokensLocked}
+                  disabled={tokenLocked}
                 />
               </span>
             </Tooltip>
           ) : (
             <Input.Password
-              value={appToken}
-              onChange={handleAppTokenChange}
-              placeholder={tokensLocked || pluginStatus?.hasToken ? '••••••••••••••••' : 'xapp-...'}
+              value={botToken}
+              onChange={handleTokenChange}
+              placeholder={tokenLocked || pluginStatus?.hasToken ? '••••••••••••••••' : 'MTAx...'}
               style={{ width: 260 }}
               visibilityToggle
-              disabled={tokensLocked}
+              disabled={tokenLocked}
             />
           )}
-          {tokensLocked ? (
+          {tokenLocked ? (
             <Tooltip
               content={t(
                 'settings.assistant.tokenLocked',
@@ -429,13 +362,13 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
               )}
             >
               <span>
-                <Button type='outline' loading={testLoading} onClick={handleTestConnection} disabled={tokensLocked}>
+                <Button type='outline' loading={testLoading} onClick={handleTestConnection} disabled={tokenLocked}>
                   {t('settings.assistant.testConnection', 'Test')}
                 </Button>
               </span>
             </Tooltip>
           ) : (
-            <Button type='outline' loading={testLoading} onClick={handleTestConnection} disabled={tokensLocked}>
+            <Button type='outline' loading={testLoading} onClick={handleTestConnection} disabled={tokenLocked}>
               {t('settings.assistant.testConnection', 'Test')}
             </Button>
           )}
@@ -448,7 +381,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
           label={t('settings.assistant.name', 'Assistant')}
           description={
             <div className='flex flex-col gap-4px'>
-              <span>{t('settings.assistant.agentDescSlack', 'Used for Slack conversations')}</span>
+              <span>{t('settings.assistant.agentDescDiscord', 'Used for Discord conversations')}</span>
               {hasBrokenSavedAssistant && (
                 <span className='text-orange-6'>
                   {t(
@@ -509,7 +442,7 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
       {/* Default Model Selection */}
       <PreferenceRow
         label={t('settings.assistant.defaultModel', 'Default Model')}
-        description={t('settings.assistant.defaultModelDescSlack', 'Model used for Slack conversations')}
+        description={t('settings.assistant.defaultModelDescDiscord', 'Model used for Discord conversations')}
       >
         <GoogleModelSelector
           selection={showModelSelector ? modelSelection : undefined}
@@ -530,7 +463,10 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
           <div className='text-14px text-t-secondary space-y-8px'>
             <p className='m-0'>
               <strong>1.</strong>{' '}
-              {t('settings.assistant.slackStep1', 'Open Slack and mention your bot in a channel, or send it a DM')}
+              {t(
+                'settings.assistant.discordStep1',
+                'Open Discord and mention your bot in a server channel, or send it a DM'
+              )}
               {pluginStatus.botUsername && (
                 <span className='ms-4px'>
                   <code className='bg-fill-2 px-6px py-2px rd-4px'>@{pluginStatus.botUsername}</code>
@@ -539,18 +475,21 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
             </p>
             <p className='m-0'>
               <strong>2.</strong>{' '}
-              {t('settings.assistant.slackStep2', 'Send any message or @mention to initiate pairing')}
+              {t('settings.assistant.discordStep2', 'Send any message or @mention to initiate pairing')}
             </p>
             <p className='m-0'>
               <strong>3.</strong>{' '}
               {t(
-                'settings.assistant.slackStep3',
+                'settings.assistant.discordStep3',
                 'A pairing request will appear below. Click "Approve" to authorize the user.'
               )}
             </p>
             <p className='m-0'>
               <strong>4.</strong>{' '}
-              {t('settings.assistant.slackStep4', 'Once approved, you can start chatting with the assistant in Slack!')}
+              {t(
+                'settings.assistant.discordStep4',
+                'Once approved, you can start chatting with the assistant in Discord!'
+              )}
             </p>
           </div>
         </div>
@@ -688,4 +627,4 @@ const SlackConfigForm: React.FC<SlackConfigFormProps> = ({
   );
 };
 
-export default SlackConfigForm;
+export default DiscordConfigForm;
