+++ packages/desktop/src/renderer/components/settings/SettingsModal/contents/channels/platforms/TelegramConfigForm.tsx
@@ -8,7 +8,6 @@ import type { IChannelPairingRequest, IChannelPluginStatus, IChannelUser } from 
 import { assistants, channel } from '@/common/adapter/ipcBridge';
 import { isAionrsAssistant, type Assistant } from '@/common/types/agent/assistantTypes';
 import { resolveLocaleKey } from '@/common/utils';
-import { openExternalUrl } from '@/renderer/utils/platform';
 import { resolveAssistantName } from '@/renderer/utils/model/assistantDisplay';
 import { formatDateTime } from '@/renderer/services/i18n/format';
 import GoogleModelSelector from '@/renderer/pages/conversation/platforms/gemini/GoogleModelSelector';
@@ -21,7 +20,7 @@ import {
   buildChannelAssistantBinding,
   getDefaultChannelAssistant,
   resolveChannelAssistantSelection,
-} from './assistantBinding';
+} from '../assistantBinding';
 
 /**
  * Preference row component
@@ -30,16 +29,12 @@ const PreferenceRow: React.FC<{
   label: string;
   description?: React.ReactNode;
   extra?: React.ReactNode;
-  required?: boolean;
   children: React.ReactNode;
-}> = ({ label, description, extra, required, children }) => (
+}> = ({ label, description, extra, children }) => (
   <div className='flex items-center justify-between gap-24px py-12px'>
     <div className='flex-1'>
       <div className='flex items-center gap-8px'>
-        <span className='text-14px text-t-primary'>
-          {label}
-          {required && <span className='text-red-500 ms-2px'>*</span>}
-        </span>
+        <span className='text-14px text-t-primary'>{label}</span>
         {extra}
       </div>
       {description && <div className='text-12px text-t-tertiary mt-2px'>{description}</div>}
@@ -58,30 +53,32 @@ const SectionHeader: React.FC<{ title: string; action?: React.ReactNode }> = ({ 
   </div>
 );
 
-interface DingTalkConfigFormProps {
+interface TelegramConfigFormProps {
   pluginStatus: IChannelPluginStatus | null;
   modelSelection: GoogleModelSelection;
   onStatusChange: (status: IChannelPluginStatus | null) => void;
+  onTokenChange?: (token: string) => void;
 }
 
-const DINGTALK_DEV_DOCS_URL = 'https://github.com/iOfficeAI/AionUi/wiki/DingTalk-Bot-Setup-Guide';
-
-const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, modelSelection, onStatusChange }) => {
+const TelegramConfigForm: React.FC<TelegramConfigFormProps> = ({
+  pluginStatus,
+  modelSelection,
+  onStatusChange,
+  onTokenChange,
+}) => {
   const { t, i18n } = useTranslation();
   const localeKey = resolveLocaleKey(i18n?.language ?? 'en-US');
 
-  // DingTalk credentials
-  const [clientId, setClientId] = useState('');
-  const [clientSecret, setClientSecret] = useState('');
-
+  const [telegramToken, setTelegramToken] = useState('');
   const [testLoading, setTestLoading] = useState(false);
-  const [_credentialsTested, setCredentialsTested] = useState(false);
-  const [touched, setTouched] = useState({ clientId: false, clientSecret: false });
+  const [tokenTested, setTokenTested] = useState(false);
+  const [testedBotUsername, setTestedBotUsername] = useState<string | null>(null);
   const [pairingLoading, setPairingLoading] = useState(false);
   const [usersLoading, setUsersLoading] = useState(false);
   const [pendingPairings, setPendingPairings] = useState<IChannelPairingRequest[]>([]);
   const [authorizedUsers, setAuthorizedUsers] = useState<IChannelUser[]>([]);
 
+  // Assistant selection (used for Telegram conversations)
   const [availableAssistants, setAvailableAssistants] = useState<Assistant[]>([]);
   const [selectedAssistant, setSelectedAssistant] = useState<Assistant | null>(null);
   const [hasBrokenSavedAssistant, setHasBrokenSavedAssistant] = useState(false);
@@ -92,10 +89,10 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
     try {
       const pairings = await channel.getPendingPairings.invoke();
       if (pairings) {
-        setPendingPairings(pairings.filter((p) => p.platformType === 'dingtalk'));
+        setPendingPairings(pairings.filter((p) => p.platformType === 'telegram'));
       }
     } catch (error) {
-      console.error('[DingTalkConfig] Failed to load pending pairings:', error);
+      console.error('[ChannelSettings] Failed to load pending pairings:', error);
     } finally {
       setPairingLoading(false);
     }
@@ -107,10 +104,10 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
     try {
       const users = await channel.getAuthorizedUsers.invoke();
       if (users) {
-        setAuthorizedUsers(users.filter((u) => u.platformType === 'dingtalk'));
+        setAuthorizedUsers(users.filter((u) => u.platformType === 'telegram'));
       }
     } catch (error) {
-      console.error('[DingTalkConfig] Failed to load authorized users:', error);
+      console.error('[ChannelSettings] Failed to load authorized users:', error);
     } finally {
       setUsersLoading(false);
     }
@@ -128,7 +125,7 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
       try {
         const [assistantList, saved] = await Promise.all([
           assistants.list.invoke(),
-          channel.getPlatformSettings.invoke({ platform: 'dingtalk' }),
+          channel.getPlatformSettings.invoke({ platform: 'telegram' }),
         ]);
 
         setAvailableAssistants(assistantList);
@@ -142,7 +139,7 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
         setHasBrokenSavedAssistant(selection.hasBrokenSavedAssistant);
         setSelectedAssistant(nextAssistant);
       } catch (error) {
-        console.error('[DingTalkConfig] Failed to load assistants:', error);
+        console.error('[TelegramConfig] Failed to load assistants:', error);
       }
     };
 
@@ -152,12 +149,12 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
   const persistSelectedAssistant = async (assistant: Assistant) => {
     try {
       await channel.setAssistantSetting.invoke({
-        platform: 'dingtalk',
+        platform: 'telegram',
         assistant: buildChannelAssistantBinding(assistant),
       });
       Message.success(t('settings.assistant.agentSwitched', 'Assistant switched successfully'));
     } catch (error) {
-      console.error('[DingTalkConfig] Failed to save assistant:', error);
+      console.error('[TelegramConfig] Failed to save assistant:', error);
       Message.error(t('common.saveFailed', 'Failed to save'));
     }
   };
@@ -165,7 +162,7 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
   // Listen for pairing requests
   useEffect(() => {
     const unsubscribe = channel.pairingRequested.on((request) => {
-      if (request.platformType !== 'dingtalk') return;
+      if (request.platformType !== 'telegram') return;
       setPendingPairings((prev) => {
         const exists = prev.some((p) => p.code === request.code);
         if (exists) return prev;
@@ -178,7 +175,6 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
   // Listen for user authorization
   useEffect(() => {
     const unsubscribe = channel.userAuthorized.on((user) => {
-      if (user.platformType !== 'dingtalk') return;
       setAuthorizedUsers((prev) => {
         const exists = prev.some((u) => u.id === user.id);
         if (exists) return prev;
@@ -189,38 +185,42 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
     return () => unsubscribe();
   }, []);
 
-  // Test DingTalk connection
+  // Test Telegram connection
   const handleTestConnection = async () => {
-    setTouched({ clientId: true, clientSecret: true });
-
-    if (!clientId.trim() || !clientSecret.trim()) {
-      Message.warning(t('settings.dingtalk.credentialsRequired', 'Please enter Client ID and Client Secret'));
+    if (!telegramToken.trim()) {
+      Message.warning(t('settings.assistant.tokenRequired', 'Please enter a bot token'));
       return;
     }
 
     setTestLoading(true);
-    setCredentialsTested(false);
+    setTokenTested(false);
+    setTestedBotUsername(null);
     try {
       // testPlugin returns { success, botUsername?, error? } directly
       const result = await channel.testPlugin.invoke({
-        plugin_id: 'dingtalk',
-        token: clientId.trim(),
-        extra_config: {
-          app_secret: clientSecret.trim(),
-        },
+        plugin_id: 'telegram',
+        token: telegramToken.trim(),
       });
 
       if (result.success) {
-        setCredentialsTested(true);
-        Message.success(t('settings.dingtalk.connectionSuccess', 'Connected to DingTalk API!'));
+        setTokenTested(true);
+        setTestedBotUsername(result.bot_username || null);
+        Message.success(
+          t('settings.assistant.connectionSuccess', {
+            defaultValue: 'Connected! Bot: @{{username}}',
+            username: result.bot_username || 'unknown',
+          })
+        );
+
+        // Auto-enable bot after successful test
         await handleAutoEnable();
       } else {
-        setCredentialsTested(false);
-        Message.error(result.error || t('settings.dingtalk.connectionFailed', 'Connection failed'));
+        setTokenTested(false);
+        Message.error(result.error || t('settings.assistant.connectionFailed', 'Connection failed'));
       }
     } catch (error: any) {
-      setCredentialsTested(false);
-      Message.error(error.message || t('settings.dingtalk.connectionFailed', 'Connection failed'));
+      setTokenTested(false);
+      Message.error(error.message || t('settings.assistant.connectionFailed', 'Connection failed'));
     } finally {
       setTestLoading(false);
     }
@@ -229,34 +229,29 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
   // Auto-enable plugin after successful test
   const handleAutoEnable = async () => {
     try {
+      // enablePlugin returns void; success if no throw
       await channel.enablePlugin.invoke({
-        plugin_id: 'dingtalk',
-        config: {
-          credentials: {
-            client_id: clientId.trim(),
-            client_secret: clientSecret.trim(),
-          },
-        },
+        plugin_id: 'telegram',
+        config: { credentials: { token: telegramToken.trim() } },
       });
 
-      Message.success(t('settings.dingtalk.pluginEnabled', 'DingTalk bot enabled'));
+      Message.success(t('settings.assistant.pluginEnabled', 'Telegram bot enabled'));
       const plugins = await channel.getPluginStatus.invoke();
       if (plugins) {
-        const dingtalkPlugin = plugins.find((p) => p.type === 'dingtalk');
-        onStatusChange(dingtalkPlugin || null);
+        const telegramPlugin = plugins.find((p) => p.type === 'telegram');
+        onStatusChange(telegramPlugin || null);
       }
     } catch (error: unknown) {
-      console.error('[DingTalkConfig] Auto-enable failed:', error);
-      Message.error(
-        (error instanceof Error ? error.message : String(error)) ||
-          t('settings.dingtalk.enableFailed', 'Failed to enable DingTalk plugin')
-      );
+      console.error('[ChannelSettings] Auto-enable failed:', error);
     }
   };
 
-  // Reset credentials tested state when credentials change
-  const handleCredentialsChange = () => {
-    setCredentialsTested(false);
+  // Reset token tested state when token changes
+  const handleTokenChange = (value: string) => {
+    setTelegramToken(value);
+    setTokenTested(false);
+    setTestedBotUsername(null);
+    onTokenChange?.(value);
   };
 
   // Approve pairing
@@ -309,7 +304,6 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
     return Math.max(0, Math.ceil((expiresAt - Date.now()) / 1000 / 60));
   };
 
-  const hasExistingUsers = authorizedUsers.length > 0;
   const showModelSelector = isAionrsAssistant(selectedAssistant);
   const assistantOptions = availableAssistants;
   const selectedAssistantName = selectedAssistant
@@ -318,142 +312,76 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
 
   return (
     <div className='flex flex-col gap-24px'>
-      {/* Client ID */}
       <PreferenceRow
-        label={t('settings.dingtalk.clientId', 'Client ID')}
-        description={
-          <span>
-            <a
-              className='text-primary hover:underline cursor-pointer text-12px'
-              href={DINGTALK_DEV_DOCS_URL}
-              onClick={(e) => {
-                e.preventDefault();
-                openExternalUrl(DINGTALK_DEV_DOCS_URL).catch(console.error);
-              }}
-            >
-              {t('settings.dingtalk.devConsoleLink', 'DingTalk Open Platform')}
-            </a>{' '}
-            {t('settings.dingtalk.clientIdDescSuffix', 'to get your Client ID')}
-          </span>
-        }
-        required
-      >
-        {hasExistingUsers ? (
-          <Tooltip
-            content={t(
-              'settings.assistant.tokenLocked',
-              'Please close the Channel and delete all authorized users before modifying'
-            )}
-          >
-            <span>
-              <Input
-                value={clientId}
-                onChange={(value) => {
-                  setClientId(value);
-                  handleCredentialsChange();
-                }}
-                onBlur={() => setTouched((prev) => ({ ...prev, clientId: true }))}
-                placeholder={hasExistingUsers || pluginStatus?.hasToken ? '••••••••••••••••' : 'dingxxxxxxxxxx'}
-                style={{ width: 240 }}
-                status={touched.clientId && !clientId.trim() && !pluginStatus?.hasToken ? 'error' : undefined}
-                disabled={hasExistingUsers}
-              />
-            </span>
-          </Tooltip>
-        ) : (
-          <Input
-            value={clientId}
-            onChange={(value) => {
-              setClientId(value);
-              handleCredentialsChange();
-            }}
-            onBlur={() => setTouched((prev) => ({ ...prev, clientId: true }))}
-            placeholder={hasExistingUsers || pluginStatus?.hasToken ? '••••••••••••••••' : 'dingxxxxxxxxxx'}
-            style={{ width: 240 }}
-            status={touched.clientId && !clientId.trim() && !pluginStatus?.hasToken ? 'error' : undefined}
-            disabled={hasExistingUsers}
-          />
+        label={t('settings.assistant.botToken', 'Bot Token')}
+        description={t(
+          'settings.assistant.botTokenDesc',
+          'Open Telegram, find @BotFather and send /newbot to get your Bot Token.'
         )}
-      </PreferenceRow>
-
-      {/* Client Secret */}
-      <PreferenceRow
-        label={t('settings.dingtalk.clientSecret', 'Client Secret')}
-        description={
-          <span>
-            <a
-              className='text-primary hover:underline cursor-pointer text-12px'
-              href={DINGTALK_DEV_DOCS_URL}
-              onClick={(e) => {
-                e.preventDefault();
-                openExternalUrl(DINGTALK_DEV_DOCS_URL).catch(console.error);
-              }}
-            >
-              {t('settings.dingtalk.devConsoleLink', 'DingTalk Open Platform')}
-            </a>{' '}
-            {t('settings.dingtalk.clientSecretDescSuffix', 'to get Client Secret')}
-          </span>
-        }
-        required
       >
-        {hasExistingUsers ? (
-          <Tooltip
-            content={t(
-              'settings.assistant.tokenLocked',
-              'Please close the Channel and delete all authorized users before modifying'
-            )}
-          >
-            <span>
-              <Input.Password
-                value={clientSecret}
-                onChange={(value) => {
-                  setClientSecret(value);
-                  handleCredentialsChange();
-                }}
-                onBlur={() => setTouched((prev) => ({ ...prev, clientSecret: true }))}
-                placeholder={hasExistingUsers || pluginStatus?.hasToken ? '••••••••••••••••' : 'xxxxxxxxxxxxxxxxxx'}
-                style={{ width: 240 }}
-                status={touched.clientSecret && !clientSecret.trim() && !pluginStatus?.hasToken ? 'error' : undefined}
-                visibilityToggle
-                disabled={hasExistingUsers}
-              />
-            </span>
-          </Tooltip>
-        ) : (
-          <Input.Password
-            value={clientSecret}
-            onChange={(value) => {
-              setClientSecret(value);
-              handleCredentialsChange();
-            }}
-            onBlur={() => setTouched((prev) => ({ ...prev, clientSecret: true }))}
-            placeholder={hasExistingUsers || pluginStatus?.hasToken ? '••••••••••••••••' : 'xxxxxxxxxxxxxxxxxx'}
-            style={{ width: 240 }}
-            status={touched.clientSecret && !clientSecret.trim() && !pluginStatus?.hasToken ? 'error' : undefined}
-            visibilityToggle
-            disabled={hasExistingUsers}
-          />
-        )}
-      </PreferenceRow>
-
-      {/* Test Connection Button */}
-      {!hasExistingUsers && !pluginStatus?.connected && (
-        <div className='flex justify-end'>
-          {pluginStatus?.hasToken && !clientId.trim() && !clientSecret.trim() ? (
-            <span className='text-12px text-t-tertiary me-12px self-center'>
-              {t('settings.dingtalk.credentialsSaved', 'Credentials already configured. Enter new values to update.')}
-            </span>
-          ) : null}
-          <Button
-            type='primary'
-            loading={testLoading}
-            onClick={handleTestConnection}
-            disabled={pluginStatus?.hasToken && !clientId.trim() && !clientSecret.trim()}
-          >
-            {t('settings.dingtalk.testAndConnect', 'Test & Connect')}
-          </Button>
+        <div className='flex items-center gap-8px'>
+          {authorizedUsers.length > 0 ? (
+            <Tooltip
+              content={t(
+                'settings.assistant.tokenLocked',
+                'Please close the Channel and delete all authorized users before modifying the configuration'
+              )}
+            >
+              <span>
+                <Input.Password
+                  value={telegramToken}
+                  onChange={handleTokenChange}
+                  placeholder={
+                    authorizedUsers.length > 0 || pluginStatus?.hasToken ? '••••••••••••••••' : '123456:ABC-DEF...'
+                  }
+                  style={{ width: 240 }}
+                  visibilityToggle
+                  disabled={authorizedUsers.length > 0}
+                />
+              </span>
+            </Tooltip>
+          ) : (
+            <Input.Password
+              value={telegramToken}
+              onChange={handleTokenChange}
+              placeholder={
+                authorizedUsers.length > 0 || pluginStatus?.hasToken ? '••••••••••••••••' : '123456:ABC-DEF...'
+              }
+              style={{ width: 240 }}
+              visibilityToggle
+              disabled={authorizedUsers.length > 0}
+            />
+          )}
+          {authorizedUsers.length > 0 ? (
+            <Tooltip
+              content={t(
+                'settings.assistant.tokenLocked',
+                'Please close the Channel and delete all authorized users before modifying the configuration'
+              )}
+            >
+              <span>
+                <Button
+                  type='outline'
+                  loading={testLoading}
+                  onClick={handleTestConnection}
+                  disabled={authorizedUsers.length > 0}
+                >
+                  {t('settings.assistant.testConnection', 'Test')}
+                </Button>
+              </span>
+            </Tooltip>
+          ) : (
+            <Button
+              type='outline'
+              loading={testLoading}
+              onClick={handleTestConnection}
+              disabled={authorizedUsers.length > 0}
+            >
+              {t('settings.assistant.testConnection', 'Test')}
+            </Button>
+          )}
         </div>
-      )}
+      </PreferenceRow>
 
       {/* Assistant Selection */}
       <div className='flex flex-col gap-8px'>
@@ -461,7 +389,7 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
           label={t('settings.assistant.name', 'Assistant')}
           description={
             <div className='flex flex-col gap-4px'>
-              <span>{t('settings.dingtalk.agentDesc', 'Used for DingTalk conversations')}</span>
+              <span>{t('settings.assistant.agentDescTelegram', 'Used for Telegram conversations')}</span>
               {hasBrokenSavedAssistant && (
                 <span className='text-orange-6'>
                   {t(
@@ -484,9 +412,8 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
                     <Menu.Item
                       key={assistant.id}
                       onClick={() => {
-                        if (assistant.id === selectedAssistant?.id) {
-                          return;
-                        }
+                        if (assistant.id === selectedAssistant?.id) return;
+
                         setHasBrokenSavedAssistant(false);
                         setSelectedAssistant(assistant);
                         void persistSelectedAssistant(assistant);
@@ -522,75 +449,57 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
 
       {/* Default Model Selection */}
       <PreferenceRow
-        label={t('settings.assistant.defaultModel', 'Model')}
-        description={t('settings.dingtalk.defaultModelDesc', 'Model used for conversations handled by this assistant')}
+        label={t('settings.assistant.defaultModel', 'Default Model')}
+        description={t(
+          'settings.assistant.defaultModelDesc',
+          'Model used for Telegram conversations handled by this assistant'
+        )}
       >
         <GoogleModelSelector
           selection={showModelSelector ? modelSelection : undefined}
           disabled={!showModelSelector}
           label={
-            !showModelSelector ? t('settings.assistant.autoFollowCliModel', 'Auto-follow CLI runtime model') : undefined
+            !showModelSelector
+              ? t('settings.assistant.autoFollowCliModel', 'Automatically follow the model when CLI is running')
+              : undefined
           }
           variant='settings'
         />
       </PreferenceRow>
 
-      {/* Connection Status */}
-      {pluginStatus?.enabled && authorizedUsers.length === 0 && (
-        <div
-          className={`rd-12px p-16px border ${pluginStatus?.connected ? 'bg-green-50 dark:bg-green-900/20 border-green-200 dark:border-green-800' : pluginStatus?.error ? 'bg-red-50 dark:bg-red-900/20 border-red-200 dark:border-red-800' : 'bg-yellow-50 dark:bg-yellow-900/20 border-yellow-200 dark:border-yellow-800'}`}
-        >
-          <SectionHeader
-            title={t('settings.dingtalk.connectionStatus', 'Connection Status')}
-            action={
-              <span
-                className={`text-12px px-8px py-2px rd-4px ${pluginStatus?.connected ? 'bg-green-100 text-green-700 dark:bg-green-900 dark:text-green-300' : pluginStatus?.error ? 'bg-red-100 text-red-700 dark:bg-red-900 dark:text-red-300' : 'bg-yellow-100 text-yellow-700 dark:bg-yellow-900 dark:text-yellow-300'}`}
-              >
-                {pluginStatus?.connected
-                  ? t('settings.dingtalk.statusConnected', 'Connected')
-                  : pluginStatus?.error
-                    ? t('settings.dingtalk.statusError', 'Error')
-                    : t('settings.dingtalk.statusConnecting', 'Connecting...')}
-              </span>
-            }
-          />
-          {pluginStatus?.error && (
-            <div className='text-14px text-red-600 dark:text-red-400 mb-12px'>{pluginStatus.error}</div>
-          )}
-          {pluginStatus?.connected && (
-            <div className='text-14px text-t-secondary space-y-8px'>
-              <p className='m-0 font-500'>{t('settings.assistant.nextSteps', 'Next Steps')}:</p>
-              <p className='m-0'>
-                <strong>1.</strong> {t('settings.dingtalk.step1', 'Open DingTalk and find your bot application')}
-              </p>
-              <p className='m-0'>
-                <strong>2.</strong> {t('settings.dingtalk.step2', 'Send any message to initiate pairing')}
-              </p>
-              <p className='m-0'>
-                <strong>3.</strong>{' '}
-                {t(
-                  'settings.dingtalk.step3',
-                  'A pairing request will appear below. Click "Approve" to authorize the user.'
-                )}
-              </p>
-              <p className='m-0'>
-                <strong>4.</strong>{' '}
-                {t(
-                  'settings.dingtalk.step4',
-                  'Once approved, you can start chatting with the AI assistant through DingTalk!'
-                )}
-              </p>
-            </div>
-          )}
-          {!pluginStatus?.connected && !pluginStatus?.error && (
-            <div className='text-14px text-t-secondary'>
-              {t('settings.dingtalk.waitingConnection', 'Connection is being established. Please wait...')}
-            </div>
-          )}
+      {/* Next Steps Guide - show when bot is enabled and no authorized users yet */}
+      {pluginStatus?.enabled && pluginStatus?.connected && authorizedUsers.length === 0 && (
+        <div className='bg-blue-50 dark:bg-blue-900/20 rd-12px p-16px border border-blue-200 dark:border-blue-800'>
+          <SectionHeader title={t('settings.assistant.nextSteps', 'Next Steps')} />
+          <div className='text-14px text-t-secondary space-y-8px'>
+            <p className='m-0'>
+              <strong>1.</strong> {t('settings.assistant.step1', 'Open Telegram and search for your bot')}
+              {pluginStatus.botUsername && (
+                <span className='ms-4px'>
+                  <code className='bg-fill-2 px-6px py-2px rd-4px'>@{pluginStatus.botUsername}</code>
+                </span>
+              )}
+            </p>
+            <p className='m-0'>
+              <strong>2.</strong>{' '}
+              {t('settings.assistant.step2', 'Send any message or click /start to initiate pairing')}
+            </p>
+            <p className='m-0'>
+              <strong>3.</strong>{' '}
+              {t(
+                'settings.assistant.step3',
+                'A pairing request will appear below. Click "Approve" to authorize the user.'
+              )}
+            </p>
+            <p className='m-0'>
+              <strong>4.</strong>{' '}
+              {t('settings.assistant.step4', 'Once approved, you can start chatting with Gemini through Telegram!')}
+            </p>
+          </div>
         </div>
       )}
 
-      {/* Pending Pairings */}
+      {/* Pending Pairings - show when bot is enabled and no authorized users yet */}
       {pluginStatus?.enabled && authorizedUsers.length === 0 && (
         <div className='bg-fill-1 rd-12px pt-16px pe-16px pb-16px ps-0'>
           <SectionHeader
@@ -603,7 +512,7 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
                 loading={pairingLoading}
                 onClick={loadPendingPairings}
               >
-                {t('conversation.workspace.refresh', 'Refresh')}
+                {t('common.refresh', 'Refresh')}
               </Button>
             }
           />
@@ -665,7 +574,7 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
         </div>
       )}
 
-      {/* Authorized Users */}
+      {/* Authorized Users - show when there are authorized users */}
       {authorizedUsers.length > 0 && (
         <div className='bg-fill-1 rd-12px pt-16px pe-16px pb-16px ps-0'>
           <SectionHeader
@@ -720,4 +629,4 @@ const DingTalkConfigForm: React.FC<DingTalkConfigFormProps> = ({ pluginStatus, m
   );
 };
 
-export default DingTalkConfigForm;
+export default TelegramConfigForm;
