+++ packages/desktop/src/renderer/services/i18n/i18n-keys.d.ts
@@ -2113,10 +2113,12 @@ export type I18nKey =
   | 'settings.lightMode'
   | 'settings.logDir'
   | 'settings.mattermost.accessToken'
+  | 'settings.mattermost.accessTokenPlaceholder'
   | 'settings.mattermost.accessTokenRequired'
   | 'settings.mattermost.agentDesc'
   | 'settings.mattermost.allowedChannelIds'
   | 'settings.mattermost.allowedChannelIdsDesc'
+  | 'settings.mattermost.channelIdsPlaceholder'
   | 'settings.mattermost.configureFirst'
   | 'settings.mattermost.connectionFailed'
   | 'settings.mattermost.connectionStatus'
@@ -2128,7 +2130,9 @@ export type I18nKey =
   | 'settings.mattermost.pluginEnabled'
   | 'settings.mattermost.replyInThread'
   | 'settings.mattermost.saveAndEnable'
+  | 'settings.mattermost.savedTokenPlaceholder'
   | 'settings.mattermost.serverUrl'
+  | 'settings.mattermost.serverUrlPlaceholder'
   | 'settings.mattermost.serverUrlRequired'
   | 'settings.mattermost.statusConnected'
   | 'settings.mattermost.statusConnecting'
