+++ packages/desktop/src/renderer/services/i18n/locales/es-ES/settings.json
@@ -1417,6 +1417,10 @@
   "fontSizeMarkdown": "Tamaño de texto Markdown",
   "fontSizeCode": "Tamaño de fuente del código",
   "mattermost": {
+    "serverUrlPlaceholder": "https://mattermost.example.com",
+    "accessTokenPlaceholder": "MMAUTHTOKEN...",
+    "savedTokenPlaceholder": "••••••••••••••••",
+    "channelIdsPlaceholder": "channel_id_1, channel_id_2",
     "serverUrl": "Mattermost Server URL",
     "accessToken": "Access Token",
     "allowedChannelIds": "Allowed Channel IDs",
