+++ packages/desktop/src/renderer/services/i18n/locales/ko-KR/settings.json
@@ -1419,6 +1419,10 @@
   "fontSizeMarkdown": "Markdown 텍스트 크기",
   "fontSizeCode": "코드 글꼴 크기",
   "mattermost": {
+    "serverUrlPlaceholder": "https://mattermost.example.com",
+    "accessTokenPlaceholder": "MMAUTHTOKEN...",
+    "savedTokenPlaceholder": "••••••••••••••••",
+    "channelIdsPlaceholder": "channel_id_1, channel_id_2",
     "serverUrl": "Mattermost Server URL",
     "accessToken": "Access Token",
     "allowedChannelIds": "Allowed Channel IDs",
