+++ packages/desktop/src/renderer/services/i18n/locales/pt-BR/settings.json
@@ -1420,6 +1420,10 @@
   "fontSizeMarkdown": "Tamanho do texto Markdown",
   "fontSizeCode": "Tamanho da fonte do código",
   "mattermost": {
+    "serverUrlPlaceholder": "https://mattermost.example.com",
+    "accessTokenPlaceholder": "MMAUTHTOKEN...",
+    "savedTokenPlaceholder": "••••••••••••••••",
+    "channelIdsPlaceholder": "channel_id_1, channel_id_2",
     "serverUrl": "Mattermost Server URL",
     "accessToken": "Access Token",
     "allowedChannelIds": "Allowed Channel IDs",
