+++ packages/desktop/src/renderer/services/i18n/locales/ru-RU/settings.json
@@ -1431,6 +1431,10 @@
   "fontSizeMarkdown": "Размер текста Markdown",
   "fontSizeCode": "Размер шрифта кода",
   "mattermost": {
+    "serverUrlPlaceholder": "https://mattermost.example.com",
+    "accessTokenPlaceholder": "MMAUTHTOKEN...",
+    "savedTokenPlaceholder": "••••••••••••••••",
+    "channelIdsPlaceholder": "channel_id_1, channel_id_2",
     "serverUrl": "Mattermost Server URL",
     "accessToken": "Access Token",
     "allowedChannelIds": "Allowed Channel IDs",
