+++ packages/desktop/src/renderer/services/i18n/locales/tr-TR/settings.json
@@ -1417,6 +1417,10 @@
   "fontSizeMarkdown": "Markdown metin boyutu",
   "fontSizeCode": "Kod yazı tipi boyutu",
   "mattermost": {
+    "serverUrlPlaceholder": "https://mattermost.example.com",
+    "accessTokenPlaceholder": "MMAUTHTOKEN...",
+    "savedTokenPlaceholder": "••••••••••••••••",
+    "channelIdsPlaceholder": "channel_id_1, channel_id_2",
     "serverUrl": "Mattermost Server URL",
     "accessToken": "Access Token",
     "allowedChannelIds": "Allowed Channel IDs",
