+++ packages/desktop/src/renderer/services/i18n/locales/uk-UA/settings.json
@@ -1432,6 +1432,10 @@
   "fontSizeMarkdown": "Розмір тексту Markdown",
   "fontSizeCode": "Розмір шрифту коду",
   "mattermost": {
+    "serverUrlPlaceholder": "https://mattermost.example.com",
+    "accessTokenPlaceholder": "MMAUTHTOKEN...",
+    "savedTokenPlaceholder": "••••••••••••••••",
+    "channelIdsPlaceholder": "channel_id_1, channel_id_2",
     "serverUrl": "Mattermost Server URL",
     "accessToken": "Access Token",
     "allowedChannelIds": "Allowed Channel IDs",
