+++ tests/unit/renderer/channelSettings/assistantBinding.test.ts
